karrot_api/requirements.txt
horey.configuration_policy>=1.0.0
horey.h_logger>=1.0.0
horey.selenium_api>=1.0.0
webdriver_manager